#17 Moved the implementation classes into the respective interface files.
