+++ cml-compiler/cml-io/src/main/java/cml/io/Console.java
@@ -9,3 +9,13 @@ public interface Console
         return new ConsoleImpl();
     }
 }
+
+@SuppressWarnings("UseOfSystemOutOrSystemErr")
+class ConsoleImpl implements Console
+{
+    @Override
+    public void println(final String message, final Object... args)
+    {
+        System.out.println(String.format(message, args));
+    }
+}
