--- cml-compiler/cml-io/src/main/java/cml/io/ConsoleImpl.java
@@ -1,10 +1,0 @@
-package cml.io;
-
-class ConsoleImpl implements Console
-{
-    @Override
-    public void println(final String message, final Object... args)
-    {
-        System.out.println(String.format(message, args));
-    }
-}
